Tracks a face.
Takes 10 frames of images from an attached webcam. After that, the program draws a bounding box around each face that is detected in each image. Once image has been analyzed, it plays back a greyscale version of each picture in order, essentially making a simple 10 frame video with motion tracking

Requires the cam2ImgArray function, the frameToPicArray function, and the webcam add-on for Matlab.